requirements.txt
requests
lxml
tqdm
pillow
